remove powdr-executor dependency from backend-utils and plonky3
