+++ Cargo.toml
@@ -29,6 +29,7 @@ members = [
     "riscv-syscalls",
     "schemas",
     "backend-utils",
+    "executor-utils",
 ]
 
 exclude = [ "riscv-runtime" ]
@@ -51,6 +52,7 @@ powdr-analysis = { path = "./analysis", version = "0.1.3" }
 powdr-backend = { path = "./backend", version = "0.1.3" }
 powdr-backend-utils = { path = "./backend-utils", version = "0.1.3" }
 powdr-executor = { path = "./executor", version = "0.1.3" }
+powdr-executor-utils = { path = "./executor-utils", version = "0.1.3" }
 powdr-importer = { path = "./importer", version = "0.1.3" }
 powdr-jit-compiler = { path = "./jit-compiler", version = "0.1.3" }
 powdr-linker = { path = "./linker", version = "0.1.3" }
