+++ backend-utils/Cargo.toml
@@ -11,6 +11,6 @@ powdr-pil-analyzer.workspace = true
 powdr-parser.workspace = true
 powdr-ast.workspace = true
 powdr-number.workspace = true
-powdr-executor.workspace = true
+powdr-executor-utils.workspace = true
 log = "0.4.22"
 itertools = "0.13.0"
