+++ executor/src/constant_evaluator/data_structures.rs
@@ -1,36 +1,4 @@
-use powdr_number::DegreeType;
-use serde::{Deserialize, Serialize};
-use std::collections::{BTreeMap, BTreeSet};
-
-#[derive(Serialize, Deserialize)]
-pub struct VariablySizedColumn<F> {
-    column_by_size: BTreeMap<DegreeType, Vec<F>>,
-}
-
-#[derive(Debug)]
-pub struct HasMultipleSizesError;
-
-impl<F> VariablySizedColumn<F> {
-    /// Create a view where each column has a single size. Fails if any column has multiple sizes.
-    pub fn get_uniquely_sized(&self) -> Result<&Vec<F>, HasMultipleSizesError> {
-        if self.column_by_size.len() != 1 {
-            return Err(HasMultipleSizesError);
-        }
-        Ok(self.column_by_size.values().next().unwrap())
-    }
-
-    /// Returns the set of available sizes.
-    pub fn available_sizes(&self) -> BTreeSet<DegreeType> {
-        self.column_by_size.keys().cloned().collect()
-    }
-
-    /// Clones and returns the column with the given size.
-    pub fn get_by_size(&self, size: DegreeType) -> Option<&[F]> {
-        self.column_by_size
-            .get(&size)
-            .map(|column| column.as_slice())
-    }
-}
+pub use powdr_executor_utils::{HasMultipleSizesError, VariablySizedColumn};
 
 /// Returns all columns with their unique sizes. Fails if any column has multiple sizes.
 pub fn get_uniquely_sized<F>(
@@ -51,23 +19,4 @@ pub fn get_uniquely_sized_cloned<F: Clone>(
             .map(|(name, column)| (name, column.clone()))
             .collect()
     })
-}
-
-impl<F> From<Vec<F>> for VariablySizedColumn<F> {
-    fn from(column: Vec<F>) -> Self {
-        VariablySizedColumn {
-            column_by_size: [(column.len() as DegreeType, column)].into_iter().collect(),
-        }
-    }
-}
-
-impl<F> From<Vec<Vec<F>>> for VariablySizedColumn<F> {
-    fn from(columns: Vec<Vec<F>>) -> Self {
-        VariablySizedColumn {
-            column_by_size: columns
-                .into_iter()
-                .map(|column| (column.len() as DegreeType, column))
-                .collect(),
-        }
-    }
 }
